incluye linea base, y correcciones de atributos. Cambios en algunas lineas en base.html

+++ GPMmodulos/webapp/vistas/views_adm.py
@@ -99,17 +99,21 @@ def deleteUser(user_id):
     """Funcion encargada de eliminar un usuario del sistema"""
     user = User.query.filter_by(id=user_id).first_or_404()
     form = DeleteUserForm(obj=user, next=request.args.get('next'))
-
-    if form.validate_on_submit():
-        form.populate_obj(user)
-
-        db.session.delete(user)
-        db.session.commit()
-
-        flash('Usuario eliminado.', 'success')
+    
+    if user == current_user:
+        flash ('Usuario no eliminado. Logueado en este momento', 'error')
         return redirect(url_for('admin.users'))
-
-    return render_template('admin/deleteUser.html', user=user, form=form)
+    else:
+        if form.validate_on_submit():
+            form.populate_obj(user)
+    
+            db.session.delete(user)
+            db.session.commit()
+    
+            flash('Usuario eliminado.', 'success')
+            return redirect(url_for('admin.users'))
+    
+        return render_template('admin/deleteUser.html', user=user, form=form)
 
 
 @admin.route('/searchUser')
@@ -422,9 +426,15 @@ def desasignarUsuario(proyecto_id, user_id):
     usuariosAsignados = proyecto.usuarioPorProyecto
     
     for item in usuariosAsignados:
+        for fase in proyecto.fases:
+            if usuarioDesasignar.id == fase.lider_fase:
+                flash('El usuario es Lider de una fase, no puede ser desasignado.', 'error')
+                return redirect(url_for('admin.usuariosxproyecto', proyecto_id=proyecto.id))
+        
         if usuarioDesasignar.id == proyecto.lider_proyecto:
             flash('El usuario es Lider del proyecto, no puede ser desasignado.', 'error')
             return redirect(url_for('admin.usuariosxproyecto', proyecto_id=proyecto.id))
+        
         elif item == usuarioDesasignar:
                 proyecto.usuarioPorProyecto.remove(item)
 #                 if usuarioDesasignar.esLiderDeFase(proyecto.id):
@@ -470,8 +480,21 @@ def crearFase(proyecto_id):
     if proyecto.fases.count() < proyecto.numero_fases:
         form = CrearFaseForm(next=request.args.get('next'))
         form.lider_fase.choices=[(g.id, g.nombre) for g in proyecto.usuarioPorProyecto]
-       
-        form.numero_fase.choices=[(i+1,i+1) for i in range(proyecto.numero_fases)]
+        
+        
+        ordenesNoExistentes = []
+        ordenesExistentes = []
+        for fase in proyecto.fases:
+            ordenesExistentes.append(fase.numero_fase)
+            
+        for i in range(proyecto.numero_fases):
+            if i+1 not in ordenesExistentes:    
+                    ordenesNoExistentes.append((i+1, i+1))
+
+         
+        
+        form.numero_fase.choices=[(i,i) for i,i in ordenesNoExistentes]
+        ordenesNoExistentes = None
         if form.validate_on_submit():
             fase = Fase()
             fase.nombre = form.nombre.data
@@ -566,14 +589,18 @@ def crearAtributo(proyecto_id, tipoItem_id):
     proyecto = Proyecto.query.filter_by(id=proyecto_id).first_or_404()
     tipoItem = TipoItem.query.filter_by(id=tipoItem_id).first_or_404()
     
-    atributos = Atributo.query.all()
+    
     form = CrearAtributoForm(next=request.args.get('next'))
-    form.atributo_id.choices=[(h.id, h.nombre) for h in atributos]
+    atributos = [(1,'String'), (2,'Integer'),(3,'Date')]
+    form.tipo.choices=[(id, nombre) for id, nombre in atributos]
    
     if form.validate_on_submit():
         atributo = Atributo()
         atributo.nombre = form.nombre.data
-        atributo.valorString = form.valor.data
+        atributo.tipo = form.tipo.data
+        #if atributo.getTipo() == 'String' form.valor.data  
+            
+        atributo.setValor(form.valor.data)  
                 
         db.session.add(atributo)
         db.session.commit()
@@ -637,6 +664,24 @@ def buscarTipoItem():
     else:
         flash('Por favor, ingrese dato a buscar','error')
     return render_template('index/buscarTipoItem.html', pagination=pagination , keywords=keywords)
+
+@admin.route('/borrarTipoItem/<proyecto_id>/<tipoItem_id>', methods=['GET', 'POST'])
+@login_required
+def borrarTipoItem(proyecto_id, tipoItem_id):
+    """Funcion que permite editar un comite"""
+    tipoItem = TipoItem.query.filter_by(id=tipoItem_id).first_or_404()
+    proyecto = Proyecto.query.filter_by(id=proyecto_id).first_or_404()
+    fases = proyecto.fases
+    
+    for fase in fases:
+        if tipoItem in fase.tipoItemPorFase:
+            flash('El tipo de item se encuentra asignado a una fase.', 'error')
+            return redirect(url_for('admin.tiposItemxproyecto',proyecto_id=proyecto.id))
+    
+    db.session.delete(tipoItem)
+    db.session.commit()
+    flash('Tipo de Item eliminado.', 'success')
+    return redirect(url_for('admin.tiposItemxproyecto',proyecto_id=proyecto.id))
 
 
 ##########################################################################
